[-Wtautological-compare] comparison of unsigned expression >= 0 is always true
Closes #95.

+++ Foundation/GorgonHeap.h
@@ -155,7 +155,7 @@ namespace wustl_mm {
 		void GorgonHeap<T>::BuildHeap(vector<T> & values) {
 			this->values.clear();
 			this->values = values;
-			for(unsigned int i = values.size() / 2; i >= 0; i--) {
+			for(int i = (int)values.size() / 2; i >= 0; i--) {
 				Heapify(i);
 			}
